requirements.txt
amqp==5.1.1
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers
drf-yasg==1.21.4
factory-boy==3.2.1
Faker==16.6.0
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.3.0
greenlet==2.0.1
haversine==2.7.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==23.0
pathlib-mate==1.2.1
prettytable==3.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
Pillow==10.0.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
SQLAlchemy==1.4.46
sqlalchemy-mate==1.4.28.3
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
